fixed by mkurnikov, vitvakatu recommendations
