--- dove/tests/cmd_dove_new_default.rs
@@ -1,60 +1,0 @@
-#![cfg(test)]
-
-use std::fs::{read_to_string};
-use dove::cli::execute;
-use toml::Value;
-
-mod test_cmd_helper;
-use crate::test_cmd_helper::{
-    project_remove, project_start, project_build, set_dependencies_local_move_stdlib,
-};
-
-/// $ dove new demoproject_25
-/// $ dove build
-/// project: demoproject_25
-#[test]
-fn default() {
-    // Project name and path
-    let project_name = "demoproject_25";
-    let (base_folder, project_folder) = project_start(project_name);
-    project_remove(&project_folder);
-
-    let args = &["dove", "new", project_name];
-    let command_string: String = args.join(" ").to_string();
-    execute(args, base_folder.to_path_buf()).unwrap_or_else(|err| {
-        panic!(
-            "[COMMAND] {}\r\n[FOLDER] {}\r\n[ERROR] {}\r\n",
-            &command_string,
-            base_folder.display(),
-            err.to_string()
-        )
-    });
-
-    // Check config
-    let package = read_to_string(project_folder.clone().join("Dove.toml"))
-        .unwrap()
-        .parse::<Value>()
-        .unwrap()
-        .get("package")
-        .unwrap()
-        .clone();
-
-    assert!(
-        package
-            .get("name")
-            .unwrap()
-            .to_string()
-            .contains(project_name),
-        "Dove.toml: invalid name",
-    );
-
-    assert!(
-        package.get("dialect").unwrap().to_string().contains("pont"),
-        "Dove.toml: invalid dialect",
-    );
-
-    set_dependencies_local_move_stdlib(&project_folder);
-    // $dove build
-    project_build(&project_folder);
-    project_remove(&project_folder);
-}
