--- dove/tests/cmd_dove_new_pont.rs
@@ -1,119 +1,0 @@
-#![cfg(test)]
-
-use std::fs::{read_to_string};
-use dove::cli::execute;
-use toml::Value;
-
-mod test_cmd_helper;
-use crate::test_cmd_helper::{
-    project_start, project_remove, set_dependencies_local_move_stdlib, project_build,
-};
-
-/// $ dove new demoproject_34 -d pont
-/// $ dove build
-/// project: demoproject_34
-#[test]
-fn default() {
-    // Project name and path
-    let project_name = "demoproject_34";
-    let (base_folder, project_folder) = project_start(project_name);
-    project_remove(&project_folder);
-
-    // $ dove new demoproject_34 -d pont
-    let args = &["dove", "new", &project_name, "-d", "pont"];
-    let command_string: String = args.join(" ").to_string();
-    execute(args, base_folder.clone()).unwrap_or_else(|err| {
-        panic!(
-            "[COMMAND] {}\r\n[FOLDER] {}\r\n[ERROR] {}\r\n",
-            &command_string,
-            &base_folder.display(),
-            err.to_string()
-        )
-    });
-
-    // Check config
-    let package = read_to_string(project_folder.clone().join("Dove.toml"))
-        .unwrap()
-        .parse::<Value>()
-        .unwrap()
-        .get("package")
-        .unwrap()
-        .clone();
-
-    assert!(
-        package
-            .get("name")
-            .unwrap()
-            .to_string()
-            .contains(project_name),
-        "Dove.toml: invalid name",
-    );
-
-    assert!(
-        package.get("dialect").unwrap().to_string().contains("pont"),
-        "Dove.toml: invalid dialect",
-    );
-
-    set_dependencies_local_move_stdlib(&project_folder);
-    // $ dove build
-    project_build(&project_folder);
-    project_remove(&project_folder);
-}
-
-/// $ dove new demoproject_29 -d pont -a ###
-/// $ dove build
-/// project: demoproject_29
-#[test]
-fn with_address() {
-    // Project name and path
-    let project_name = "demoproject_29";
-    let (base_folder, project_folder) = project_start(project_name);
-    project_remove(&project_folder);
-
-    for address in &["5CdCiQzNRZXWx7wNVCVjPMzGBFpkYHe3WKrGzd6TG97vKbnv", "0x1"] {
-        // $ dove new demoproject_29 -d pont -a ###
-        let args = &["dove", "new", &project_name, "-d", "pont", "-a", address];
-        let command_string: String = args.join(" ").to_string();
-        execute(args, base_folder.clone()).unwrap_or_else(|err| {
-            panic!(
-                "[COMMAND] {}\r\n[FOLDER] {}\r\n[ERROR] {}\r\n",
-                &command_string,
-                &base_folder.display(),
-                err.to_string()
-            )
-        });
-        // Check config
-        let package = read_to_string(project_folder.clone().join("Dove.toml"))
-            .unwrap()
-            .parse::<Value>()
-            .unwrap()
-            .get("package")
-            .unwrap()
-            .clone();
-        assert!(
-            package
-                .get("name")
-                .unwrap()
-                .to_string()
-                .contains(project_name),
-            "Dove.toml: invalid name",
-        );
-        assert!(
-            package.get("dialect").unwrap().to_string().contains("pont"),
-            "Dove.toml: invalid dialect",
-        );
-        assert!(
-            package
-                .get("account_address")
-                .unwrap()
-                .to_string()
-                .contains(address),
-            "Dove.toml: invalid account_address",
-        );
-
-        set_dependencies_local_move_stdlib(&project_folder);
-        // $ dove build
-        project_build(&project_folder);
-        project_remove(&project_folder);
-    }
-}
